+++ examples/data-sources/stdlib_replace/data-source.tf
@@ -1,6 +1,6 @@
 # Return the list with beginning value replaced.
 data "stdlib_replace" "begin" {
-  list_param     = ["foo", "two", "three"]
+  list_param     = ["foo", "bar", "two", "three"]
   replace_values = ["zero", "one"]
   index          = 0
 }
@@ -8,7 +8,7 @@ data "stdlib_replace" "begin" {
 
 # Return the list with middle values replaced.
 data "stdlib_replace" "replace" {
-  list_param     = ["zero", "foo", "bar", "four", "five"]
+  list_param     = ["zero", "foo", "bar", "baz", "four", "five"]
   replace_values = ["one", "two", "three"]
   index          = 1
 }
@@ -25,8 +25,8 @@ data "stdlib_replace" "zeroed" {
 
 # Return the list with terminating values replaced.
 data "stdlib_replace" "append" {
-  list_param     = ["zero", "foo", "bar"]
+  list_param     = ["zero", "foo", "bar", "baz"]
   replace_values = ["one", "two", "three"]
-  index          = length(["zero", "foo", "bar"]) - (length(["one", "two", "three"]) - 1)
+  index          = length(["zero", "foo", "bar", "baz"]) - (length(["one", "two", "three"]))
 }
 # => ["zero", "one", "two", "three"]