+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.4.1 (Next)
 - Add `sorted` parameter to `list_index` function.
 - Add `end_index` parameter to `replace` function.
+- Fix `end_index` auto-deduction in `replace` function.
 
 ### 1.4.0
 - Add `insert`, `replace`, and `sort_list` functions.